- Move `LexerTests` tests into their dedicated files
