+++ tests/DbmlNet.Tests.Unit/CodeAnalysis/Syntax/LexerTests.cs
@@ -5,13 +5,12 @@ using System.Linq;
 
 using DbmlNet.CodeAnalysis;
 using DbmlNet.CodeAnalysis.Syntax;
-using DbmlNet.CodeAnalysis.Text;
 
 using Xunit;
 
 namespace DbmlNet.Tests.Unit.CodeAnalysis.Syntax;
 
-public class LexerTests
+public partial class LexerTests
 {
     [Fact]
     public void Lexer_Lex_EndOfFile()
@@ -41,254 +40,6 @@ public class LexerTests
         SyntaxToken token = Assert.Single(tokens);
         Assert.Equal(SyntaxKind.BadToken, token.Kind);
         Assert.False(token.IsMissing);
-    }
-
-    [Theory]
-    [InlineData("\t")]
-    [InlineData(" ")]
-    [InlineData("  ")]
-    public void Lexer_Lex_Trivia_Whitespace(string text)
-    {
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics, includeEndOfFile: true);
-
-        Assert.Empty(diagnostics);
-        SyntaxToken token = Assert.Single(tokens);
-        SyntaxTrivia trivia = Assert.Single(token.LeadingTrivia);
-        Assert.Equal(SyntaxKind.WhitespaceTrivia, trivia.Kind);
-        Assert.Equal(text, trivia.Text);
-    }
-
-    [Theory]
-    [InlineData("\r")]
-    [InlineData("\n")]
-    [InlineData("\r\n")]
-    public void Lexer_Lex_Trivia_LineBreak(string text)
-    {
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics, includeEndOfFile: true);
-
-        Assert.Empty(diagnostics);
-        SyntaxToken token = Assert.Single(tokens);
-        SyntaxTrivia trivia = Assert.Single(token.LeadingTrivia);
-        Assert.Equal(SyntaxKind.LineBreakTrivia, trivia.Kind);
-        Assert.Equal(text, trivia.Text);
-    }
-
-    [Fact]
-    public void Lexer_Lex_Trivia_SingleLineComment()
-    {
-        string text = $"// {DataGenerator.CreateRandomMultiWordString()}";
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics, includeEndOfFile: true);
-
-        Assert.Empty(diagnostics);
-        SyntaxToken token = Assert.Single(tokens);
-        SyntaxTrivia trivia = Assert.Single(token.LeadingTrivia);
-        Assert.Equal(SyntaxKind.SingleLineCommentTrivia, trivia.Kind);
-        Assert.Equal(text, trivia.Text);
-    }
-
-    [Fact]
-    public void Lexer_Lex_Trivia_MultiLineComment()
-    {
-        string text = $"""
-        /*
-            {DataGenerator.CreateRandomMultiWordString()}
-            {DataGenerator.CreateRandomMultiWordString()}
-            {DataGenerator.CreateRandomMultiWordString()}
-            {DataGenerator.CreateRandomMultiWordString()}
-        */
-        """;
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics, includeEndOfFile: true);
-
-        Assert.Empty(diagnostics);
-        SyntaxToken token = Assert.Single(tokens);
-        SyntaxTrivia trivia = Assert.Single(token.LeadingTrivia);
-        Assert.Equal(SyntaxKind.MultiLineCommentTrivia, trivia.Kind);
-        Assert.Equal(text, trivia.Text);
-    }
-
-    [Theory]
-    [InlineData("0", 0)]
-    [InlineData("1", 1)]
-    [InlineData("1234567890", 1234567890)]
-    public void Lexer_Lex_Number(string text, decimal value)
-    {
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(SyntaxKind.NumberToken, token.Kind);
-        Assert.Equal(text, token.Text);
-        Assert.IsType<decimal>(token.Value);
-        Assert.Equal(value, token.Value);
-        Assert.False(token.IsMissing, "Token should not be missing.");
-        Assert.Empty(diagnostics);
-    }
-
-    [Theory]
-    [InlineData("0.2232", 0.2232)]
-    [InlineData("12345.67890", 12345.67890)]
-    public void Lexer_Lex_Number_WithDecimals(string text, decimal value)
-    {
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(SyntaxKind.NumberToken, token.Kind);
-        Assert.Equal(text, token.Text);
-        Assert.IsType<decimal>(token.Value);
-        Assert.Equal(value, token.Value);
-        Assert.False(token.IsMissing, "Token should not be missing.");
-        Assert.Empty(diagnostics);
-    }
-
-    [Theory]
-    [InlineData("0__.0_0_", 0.00)]
-    [InlineData("2_232.2_232", 2232.2232)]
-    [InlineData("2__23__2.2__23__2____", 2232.2232)]
-    [InlineData("1__0__0_0___.__21_22_1____", 1000.21221)]
-    public void Lexer_Lex_Number_WithSeparators(string text, decimal value)
-    {
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(SyntaxKind.NumberToken, token.Kind);
-        Assert.Equal(text, token.Text);
-        Assert.IsType<decimal>(token.Value);
-        Assert.Equal(value, token.Value);
-        Assert.False(token.IsMissing, "Token should not be missing.");
-        Assert.Empty(diagnostics);
-    }
-
-    [Fact]
-    public void Lexer_Lex_Number_WhenNumberTooLarge()
-    {
-        const string text = "99999999999999999999999999999.9";
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        Diagnostic diagnostic = Assert.Single(diagnostics);
-        Assert.Equal("The number '99999999999999999999999999999.9' is too large.", diagnostic.Message);
-        Assert.True(diagnostic.IsError, "Diagnostic show be error.");
-        Assert.False(diagnostic.IsWarning, "Diagnostic show not be warning.");
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(text, token.Text);
-        Assert.Equal(SyntaxKind.NumberToken, token.Kind);
-        Assert.False(token.IsMissing, "Token should not be missing.");
-        Assert.Null(token.Value);
-    }
-
-    [Fact]
-    public void Lexer_Lex_QuotationMarksStringToken()
-    {
-        const string textValue = """
-        error: ...the "message" is "my message".
-        """;
-        const string text = """
-        "error: ...the ""message"" is ""my message""."
-        """;
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        Assert.Empty(diagnostics);
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(SyntaxKind.QuotationMarksStringToken, token.Kind);
-        Assert.Equal(text, token.Text);
-        Assert.IsType<string>(token.Value);
-        Assert.Equal(textValue, token.Value);
-        Assert.False(token.IsMissing, "Token should not be missing.");
-    }
-
-    [Fact]
-    public void Lexer_Lex_SingleQuotationMarksStringToken()
-    {
-        const string textValue = """
-        error: ...the 'message' is 'my message'.
-        """;
-        const string text = """
-        'error: ...the ''message'' is ''my message''.'
-        """;
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        Assert.Empty(diagnostics);
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(SyntaxKind.SingleQuotationMarksStringToken, token.Kind);
-        Assert.Equal(text, token.Text);
-        Assert.IsType<string>(token.Value);
-        Assert.Equal(textValue, token.Value);
-        Assert.False(token.IsMissing, "Token should not be missing.");
-    }
-
-    [Fact]
-    public void Lexer_Lex_Unterminated_QuotationMarksStringToken()
-    {
-        const string text = "\"text";
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        Diagnostic diagnostic = Assert.Single(diagnostics);
-        Assert.True(diagnostic.IsError, "Diagnostic show be error.");
-        Assert.False(diagnostic.IsWarning, "Diagnostic show not be warning.");
-        Assert.Equal(new TextSpan(0, 1), diagnostic.Location.Span);
-        Assert.Equal("Unterminated string literal.", diagnostic.Message);
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(SyntaxKind.QuotationMarksStringToken, token.Kind);
-        Assert.Equal(text, token.Text);
-    }
-
-    [Fact]
-    public void Lexer_Lex_Unterminated_SingleQuotationMarksStringToken()
-    {
-        const string text = "\'text";
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics);
-
-        Diagnostic diagnostic = Assert.Single(diagnostics);
-        Assert.True(diagnostic.IsError, "Diagnostic show be error.");
-        Assert.False(diagnostic.IsWarning, "Diagnostic show not be warning.");
-        Assert.Equal(new TextSpan(0, 1), diagnostic.Location.Span);
-        Assert.Equal("Unterminated string literal.", diagnostic.Message);
-        SyntaxToken token = Assert.Single(tokens);
-        Assert.Equal(SyntaxKind.SingleQuotationMarksStringToken, token.Kind);
-        Assert.Equal(text, token.Text);
-    }
-
-    [Fact]
-    public void Lexer_Lex_Trivia_UnterminatedMultiLineComment()
-    {
-        string text = $"""
-        /*
-            {DataGenerator.CreateRandomMultiWordString()}
-            {DataGenerator.CreateRandomMultiWordString()}
-            {DataGenerator.CreateRandomMultiWordString()}
-            {DataGenerator.CreateRandomMultiWordString()}
-        /
-        """;
-
-        ImmutableArray<SyntaxToken> tokens =
-            SyntaxTree.ParseTokens(text, out ImmutableArray<Diagnostic> diagnostics, includeEndOfFile: true);
-
-        Diagnostic diagnostic = Assert.Single(diagnostics);
-        Assert.True(diagnostic.IsError, "Diagnostic show be error.");
-        Assert.False(diagnostic.IsWarning, "Diagnostic show not be warning.");
-        Assert.Equal(new TextSpan(0, 2), diagnostic.Location.Span);
-        Assert.Equal("Unterminated multi-line comment.", diagnostic.Message);
-        SyntaxToken token = Assert.Single(tokens);
-        SyntaxTrivia trivia = Assert.Single(token.LeadingTrivia);
-        Assert.Equal(SyntaxKind.MultiLineCommentTrivia, trivia.Kind);
-        Assert.Equal(text, trivia.Text);
     }
 
     [Fact]
